requirements.txt
Django==4.2.16
djangorestframework==3.15.2
drf-spectacular==0.28.0
psycopg2-binary==2.9.10
python-decouple==3.8
python-dateutil==2.8.2
whitenoise==6.6.0
gunicorn==21.2.0
redis==5.0.1
django-redis==5.4.0
django-cors-headers==4.3.1
celery==5.3.4
django-celery-beat==2.7.0
pytest==7.4.3
pytest-django==4.7.0
factory-boy==3.3.0
requests==2.31.0